Add Maven Wrapper
Previously it was necessary to install Maven to build the project. This
was not very convenient for users who did not have Maven installed or
might have a wrong version of Maven.

This commit adds the Maven Wrapper https://github.com/takari/maven-wrapper
The Maven Wrapper will automatically install Maven for you by including
a shell script within the repository. For example, now all that is needed
to install the project is Java and the following command

  ./mvnw clean install