Implemented a Basic Action System and implemented it as good as possible into TLS-Attacker. A WorkflowTrace now consists out of Actions which are Executed by the GenericWorkflowExecutor(or its subclases). The way these Actions are executed is defined by the used ActionExecutor. Messages dont specify a ConnectionEnd anymore, since this field is now implizit in the Action. The ProtocolMessages dont change anymore. The receiveAction now carries a list of configured and actually received messages. This deprecated alot of already implemented Functions
There are still some problems Remaining:
I think the TLSActionExecutor still doesnt handle received and configured Messages correctly
There is no field for actually send messages (in case of an exception while sending)
You cant run the same WorkflowTrace from clientSide and Server side anymore. You would have to swap all send and receive Actions
If one changes the ConnectionEnd in the TLS Context, all Actions should be reversed.
Receive Messages now store the actually received Data and Keep the configured Data. There still might be some confusions in the code
DTLS Retransmits are currently disabeld. DTLS itself is buggeling my Mind. I am not sure if i understood the DTLS code all the Time and refactored it always correctly.
I sometimes had to guess what the author actually meant. The Code Currently passes the Unit Tests, but if the code has enough Tests implemented.
I marked serveral parts of the Code with TODO, when i wasnt sure what to do or if there simply is still work to be done.
I did not mitigate the TLS-Fuzzer and MitM Modules. 
I didnt mitigate the Fuzzer because i thought that it is probably obsoleted by the new EvolutionaryFuzzer and i was not sure if it worth the effort to mitigate it.
I also didnt mitigate the MitM Modules, since the current State of those modules is far off from what is needed for the Action System. With the new system in place one should probably implement a ForwardMessageAction (extending MessageAction). This would implement MitM Attacks very elegantly. But since my master thesis is running out of time i did not already implement it.