feat: Add state and multi context support
A major commit introducing support for multiple context as needed
in man-in-the-middle scenarios.
Introduces a State which basically replaces the single TlsContext
used so far. State holds the workflow trace and multiple TlsContexts.
A TlsContext should now be regarded as the representative for a single
TLS connection.

+ Refactoring to switch over to state centered program execution.
+ Refactoring to offer multi context support with fallback to a default
  context
+ Refactoring to make sure that reading config values is done from
  TlsContext rather then from default Config
+ Introduce simple ForwardAction
  * Make handlers adjustTLSContext accessible for ForwardAction

(Add new workflow types to WorkflowTypeDelegate)