Add notes about "PublicKey" in RSAClientKeyExchanges
I had a hard time figuring out that the "PublicKey" field actually
contains the encrypted premaster secret in the case of RSA.
Also, it was non-obvious to me why the "PublicKey" length is parsed
since it's not contained in the structs from RFC 5246, only in the
explaining text.

This adds some explaining notes to comments and logging output.