Add PWD_CLEAR extension from RFC 8492
The extension allows the client to send a username to the server. The
server can then use the username to identify the client and select the
correct password to use for the key exchange (not implemented yet).

The username and sending of the message can be controlled by the config
xml. Sending the extension is off by default as the ciphersuite is not
in the list of default ciphersuites. The default username is taken from
the example of the RFC's appendix.