Add TLS_ECCPWD server and client key exchange
It is now possible to do a complete handshake for the TLS_ECCPWD
ciphersuites.

Most of the example and test values were taken from RFC 8492. However it
seems like at least some of the provided data is actually incorrect, so
my own results were used in the tests instead.

The server does not currently support multiple users and always uses the
data from the config regardless of the username the client sends.